+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] WIP
+ - API documentation now uses `documentation.js` with `documentation-polyglot`
+ - `gdal.calcAsync` can now convert input data to the output type before applying the user function
+ - `gdal.calcAsync` can now call a progress callback
+
 ## [3.4.2] WIP
  - Support the ArchLinux Node.js 17 ABI - see https://github.com/nodejs/node/issues/41410
  - Fix #25, always close open Datasets when the process exits
- - `gdal.calcAsync` can now convert input data to the output type before applying the user function
 
 ## [3.4.1] 2021-01-05